Version 1.1.2
Changed the return type of the internal function lrcoef() to
long long to make it possible to compute LR coefficients
greater than 2^32.  (Thanks to Anatol Kirillov for the bug
report.)

+++ lrcoef/symfcn.h
@@ -12,7 +12,7 @@ int part_length(vector *p);
 vector *part_conjugate(vector *p);
 int part_subset(vector *p1, vector *p2);
 
-int lrcoef(vector *outer, vector *inner1, vector *inner2);
+long long lrcoef(vector *outer, vector *inner1, vector *inner2);
 
 hashtab *skew(vector *outer, vector *inner, int maxrows);
 hashtab *mult(vector *sh1, vector *sh2, int maxrows);
